fixed bug in convertUnary and fixed tests

+++ mcc/test/tac/tac_tests.cpp
@@ -60,7 +60,7 @@ TEST(Tac, DIV) {
   expectedValue.append(" = 1 / 2");
 
   EXPECT_EQ(expectedValue, tac.toString());
-  EXPECT_EQ(Type::FLOAT, tac.codeLines.back().get()->getType());
+  EXPECT_EQ(Type::INT, tac.codeLines.back().get()->getType());
   EXPECT_EQ(1, tac.codeLines.size());
 }
